Add health checks
The echo server was completely rewritten to be used as test server, it
is now called the command server. It's a HTTP server with a few simple
endpoints(commands), as of now there are the /health endpoint, that just
returns a 200 ok, the /sleep endpoint that returns a 200 ok after 2
seconds and the /echo endpoint, that received a payload and responds
with the same data sent. As a little side-side-project I tried to
implement the HTTP protocol from scratch, it's VERY far from being
compliant.

The most important feature is, of course, the health checks. Now a
background process is started when ekilibri starts, this process tries
to reach the servers in the configuration and, based on the same config,
decides which servers are healthy. The configuration consists of:
- The path that ekilibri should call to know whether the server is
  healhty or not(health_check_path).
- The number of failures to consider the server as down(max_fails).
- The time window to consider when counting the errors to deciding to
  remove a server from the list of healthy servers(fail_window).
- The timeouts to connect, send the data and read from the
  server(connection_timeout, write_timeout and read_timeout,
  respectively).

The code is documented with how this process internally works and the
data structures used.

On ekilibri side there are a few problems to solve(not really complex
ones):
- Big payloads are still a problem, but requests under 1kb are (mostly)
  fine.
- The health endpoint and 404s are returning some binary trash after, what
  should be, the end of the response.