Handle cases when there are no healthy servers
If there are no healthy servers Ekilibri will response with a
504(Gateway Timeout), given that no servers are reachable, while if the
servers are not reacheable, but still considered healthy, when the
inevitable connection error occur the response will be a 502(Bad
Gateway).

It was necessary to change the point where the parsing of the request
happens so that the request is always parsed before anything is written
to the client connection.