Connection pool (#9)
Creates a connection pool so that connections can be reused instead of
creating new connections. The pool has a fixed size, at start ekilibri
will try to establish the same number of connections in the
configuration file, errors are simply logged. The health check process
tries checks if the servers are healthy, when the server is unhealthy
all connections are dropped and the pool goes to "reconnect" state, in
this state every call to get a connection will establish a new
connection. When the server is considered healthy again, remaining
connections are dropped and new ones are established.

The command server was changed to use axum instead of a custom server,
this was done so that the tests run against a correct(already validated)
HTTP implementation, so bugs in ekilibri's implementation were caught
this way.