Reorganize tests
Create new test scenarios and adapt scenarios of load balancing to be
parametrized, given that the tests were the same for the round robin and
least connections configuration.

Also add markers on every test case so that it's possible to run tests
for a specific feature without running every test.